Speed improvement; patch by El. Text from sourceforge:
"rebuildrecentchanges" can be accelerated considerably.
I've tested the new version with the German database.
It yields exactly the same result as the old version
and is about 5 times faster.

(I know this improvement will be obsolete once the
table structure is changed so that the revisions table
contains the id of the previous revision. But who knows
when that'll happen. ;-)

de:Benutzer:El


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@2242 dd0e9695-b195-4be7-bd10-2dea1a65a6b6