Move magic ISBN, RFC, and GEO links to between replaceExternalLinks and replaceInternalLinks. This is a basic fix for the infamous bug [ 708163 ] [[ISBN #]] doesn't work correctly
ISBN, RFC links accidentally put inside [[ ]] brackets will now come up
relatively straightforwardly, with extra brackets. This is much nicer
than a horror of HTML errors that there were before; does not solve
problem with embedded <a>-in-<a> tag if you put it in a link piped
text, but that's a smaller problem.

Someday a full tokenized parser will wish this problem away into the
cornfields.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@4506 dd0e9695-b195-4be7-bd10-2dea1a65a6b6