Change some search index update code to do two regexps backwards.
These tail-anchored regexps are insanely slow. The worst case comes
when Japanese or Chinese text (ie, no word spacing) is written on
a wiki configured for Western UTF-8 mode: the Unicode characters are
expanded to hex codes and the "words" are very long paragraph-length
monstrosities. On a large page the previous regexps may take over 20
seconds *each* on a 1GHz-level processor on such pages of non-trivial
length...

The reversed versions which are consistently much faster, for instance
by a factor of 10000 for the pathological case and a factor of 10 for
more typical English-style pages.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@4899 dd0e9695-b195-4be7-bd10-2dea1a65a6b6