1) Enable upload navigation link for non-logged-in users. From a usability perspective, navigation should not be surprising with random elements popping up under certain circumstances only ("How do I upload" is a FAQ). Special:Upload will show a message "You have to be logged in" for anonymous users, so no confusion should arise.
2) New variable $wgUploadNavigationUrl for setting the URL to which
the "Upload file" link in the toolbox points. The German Wikipedia is striving
towards an approach where local uploads are enabled, but the navigation link
goes to the Commons. The link to Special:Upload will be put on a policy page
about fair use. In this way, fair use uploads are meant to be allowed as an
exception. A similar approach is likely to be used for Wikinews.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@8556 dd0e9695-b195-4be7-bd10-2dea1a65a6b6