* Fix regression in cachability of generated CSS and JS for MonoBook skin,   while avoiding clobbering of different users' cached data
Added a "Vary: Accept-Encoding, Cookie" a few weeks ago on generated CSS/JS
files to prevent different users' styles from clobbering each other in
caches. Unfortunately browsers don't seem to handle Vary well, and this
caused a lot of extra hits due to poor caching.

This is now removed, replaced with an explicit "Cache-Control: private"
or "Cache-Control: public" depending on the presence of an open session
cookie or logged-in state. This should restore the friendly caching
behavior while ensuring that no users' generated data will clobber
anyone else's.

To additionally ensure that public cached CSS doesn't clobber the
private bits, smaxage=0 is set on the URL used for logged-in views,
as already done for JS.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@16649 dd0e9695-b195-4be7-bd10-2dea1a65a6b6