Refactoring:
1) Structure objects should generally know what data they describe.
Therefore, the structure class now has an optional type parameter.
(If not provided, the structure will try to guess its type from
the passed attribute, but if there are multiple ones, it will
be anonymous.)

This makes it easier to parse Records, as each one of them
will know what kind of data it contains. 

It necessitates some changes throughout to ensure that
correctly named structures are associated with the Records.
(Still incomplete, but it should be sufficient for now, until
we completely refactor the Save & Transaction code.)

May still have bugs. 

2) Reduce globals a bit. But really, the entire OmegaWikiAttributes
must die in favor of a schema definition that is parsed into
a single global.

3) Fix a minor dataset bug in Editor.php

4) Wrap view_as=raw in <pre> and serve as HTML to make sure that
any debug messages are shown correctly.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@23656 dd0e9695-b195-4be7-bd10-2dea1a65a6b6