Don't do gzip compression if the URL path ends in .gz or .tgz This confuses Safari and triggers a download of the page, even though it's pretty clearly labeled as viewable HTML. Bad Safari! Bad!
Bug is still present in Safari 2.0.4/Mac and 3.0.2/Win.

We have had a live hack for this on Wikimedia sites for .gz files;
hadn't noticed the .tgz problem before though.

While I was in there, also made a tweak so wfGzipHandler() makes
sure the Vary: header is there whether the current agent sent us
an Accept-Encoding header or not.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@24275 dd0e9695-b195-4be7-bd10-2dea1a65a6b6