Rearrange the updaters into chronological order, instead of four separate groups by type with weird ordering within them.
This should make it a lot easier to avoid regressions due to unnoticed
dependencies, like the one we had in 1.11 which broke upgrades from 1.4 or
below, when someone didn't realize there was an ordering problem and moved
one of the updates for aesthetic grouping reasons. :) (r24608)

It also makes it easier to see when past updates were added, since they're
now grouped by version.

Extension updates are still using the 'add table', 'add field', 'add index'
grouping.

I also removed the special-case checks for whether to do updates to the user
table. This was presumably meant for shared-db situations to avoid running
updates multiple times, but there's no need for it as the updates all check
if they've been previously applied.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@31598 dd0e9695-b195-4be7-bd10-2dea1a65a6b6