Revert r37872 for now -- "(bug 2971) Standardize format of links to hist/diff in Special:Contributions."
This appears to have changed "(hist) (diff)" to "(diff; hist)". I'm a bit leery of callign that "standardization" as it changes it from looking like regular Recent Chagnes to looking like enhanced Recent Changes.
Might be good to do a look around at other things, which use for example commas or pipes in parens as well, and figure out which is best to stick with.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@37898 dd0e9695-b195-4be7-bd10-2dea1a65a6b6