Instead of last commit, use $wgExemptFromUserRobotsControl to control __INDEX__/__NOINDEX__.  This defaults to null, which means $wgContentNamespaces.  Pages whose namespaces are in the array will be unaffected by __INDEX__/__NOINDEX__.  This will prevent serious vandalism, so the feature is again enabled by default (with content namespaces exempted).
As was pointed out, it's probably better anyway that the vandal noindex a page and have nothing turn up in searches if the search engine arrives at that exact time, than to have the vandal replace the page with "ARTICLE SUBJECT IS A POOPY-HEAD" and have that turn up in searches if the search engine arrives at that exact time.  :)  At any rate, this should solve the issue.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@38004 dd0e9695-b195-4be7-bd10-2dea1a65a6b6