Revert r38221, 38238 -- "Add new parser function {{apiurl}}. Also, add new global $wgApiScript because manually constructing the api script's path seemed like a bad idea."
I don't really like this in general; the API isn't meant for the UI and there should be little to no call to link to it from body content.

Additionally, I believe we're trying to move all new parser functions to the convention of using the # prefix to avoid conflict with the template namespace.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@38247 dd0e9695-b195-4be7-bd10-2dea1a65a6b6