Fix regression from r37834: accesskey tooltip hint should be given for the minor edit and watch labels on the edit page.  This code could still use refactoring, but that's a separate issue.  :)
I relied on the fact that $n is usable to catch back-references in JavaScript replacement text -- this seems to work in latest-ish versions of Firefox, Konqueror, and Opera, so I'm assuming it works everywhere.  ;)  As usual, the Web gives tons of dodgy-looking cookbooks with no useful info whenever I try to search for anything JavaScript-related.

Since I'm backporting this to 1.13 as a regression fix, this fix causes no change from 1.12 or 1.13 and so I'm not putting it in the release notes.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@39287 dd0e9695-b195-4be7-bd10-2dea1a65a6b6