Part two of allowing Special:Export to work on very long page histories...
Part 1 was disabling the output buffers, allowing us to stream output continuously.

Part 2, here, is to use an unbuffered query to load the page/revision data, as we do for the command-line scripts. (This is done in what Tim assures me is a portable, database-neutral way, unlike all the maint scripts which still use old hacks.) This lets us actually stream the *input* -- otherwise streaming the output is darn useless!

Part 3 will be fixing up things like the max execution time. :)


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@39676 dd0e9695-b195-4be7-bd10-2dea1a65a6b6