Redirect on login instead of showing a message
This saves a click.  A "You have successfully logged in" message adds no extra information, it just gets in the way -- especially if you got to the login page from a permissions error.  The welcome message is still displayed on account creation.

This slightly modifies the semantics of a hook.  It also makes the loginsuccess message unused, so I've deleted it from the maintenance files and also MessagesEn.  I assume it will automatically vanish from the other message files, or at least that there's some tool to remove it that I don't know about that's better than sed.

This was prompted by discussion on bug 15484.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@40621 dd0e9695-b195-4be7-bd10-2dea1a65a6b6