Make 40621 not break CentralAuth
If $injected_html is provided, then the redirect page on login will still be shown.  This is needed for auto-login in CentralAuth.  I can't test that the fix works, though, because I don't have CentralAuth installed.  Unfortunately, this means Wikipedia won't benefit from one less click, but I can't think of an easy way around that.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@40623 dd0e9695-b195-4be7-bd10-2dea1a65a6b6