This extension is a re-working of Evan Prodromou's Rdf module -- undertaken at Evan's suggestion -- and using the Redland RDF libraries.
The main design goal was to produce an internal PHP API which would be
somewhat easier for developers to use.  Special care was given to the way
Vocabularies, queries, and storage are handled.

I made Evan's use of the builder pattern more explicit, and extended it to
function both for data storage and retrieval.

This package requires both the redland binaries from Dave Beckett and the
LibRDF OOPHP wrapper from David Shea.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@40642 dd0e9695-b195-4be7-bd10-2dea1a65a6b6