Back out r42188 for now "* Disable this entirely unless curl is available so we don't blow up. Needs a-fixin * Move the curl calls to http::request() so we can at least start to make this work for the file_get_contents() fallback"
Sticking CURL-specific options in the Http::get call seems unwise as that ties you to a specific backend, assuming Http::get won't change. Further, it seems that if it does use the backend, it just plain won't work -- the returned data will be thrown away instead of saved to the target file.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@42238 dd0e9695-b195-4be7-bd10-2dea1a65a6b6