Enable \cancel and \cancelto in texvc
This works on Hardy, but may break all math support in the event that
some platform doesn't have support for \cancel.  I don't know much
OCaml or LaTeX, so I'm not sure how to make it import the package only
if it exists or something similarly useful.

If this gets reverted, though, at least I'll have committed it and
won't have to keep the diff in git stash forever.  :)

Clearly this requires a texvc recompile in any event.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@44790 dd0e9695-b195-4be7-bd10-2dea1a65a6b6