Allow exempting domain names from rel="nofollow"
This introduces a new configuration option, $wgNoFollowDomainExceptions.
By default this is an empty array; perhaps it should be null by default
and initialize to something extracted from $wgServer.  An appropriate
value for Wikimedia would be something like:

$wgNoFollowDomainExceptions = array( 'wikipedia.org', 'wiktionary.org',
'wikibooks.org', ... );

It's fairly silly that we're nofollowing links to our own sites.  :)

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@46078 dd0e9695-b195-4be7-bd10-2dea1a65a6b6