Fix the bug reported on Village Pump with some articles missing in search index.
Not sure what *exactly* happens when a page is moved over a redirect, but there seems to
be a missing OAI entry for the update/deletion of the redirect over which the page is moved. 
As a consequence, lucene "links" index gets out of date and thinks the new page is a redirect
and thus doesn't index it. It kinds sucks to have two sets of keys there but both are needed
since we might not always have ns:title (i.e. on page deletions) and we usually want to 
retrieve information for ns:title (i.e. when doing link analysis) without worring about page_id.

Fixed by explicitely cleaning up all "links" entries for both ns:title and page_id. 
Next time someone edits the missing pages they will be re-added to the search index. 


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/branches/lucene-search-2.1@48139 dd0e9695-b195-4be7-bd10-2dea1a65a6b6