Prepping for site default overrides of copyright/TOS statements for Wikimedia sites with updated CC-BY-SA license... Messages now available in WikimediaMessages extension for localization: * wikimedia-copyright - page footer * wikimedia-copyrightwarning - edit page, above summary/save button * wikimedia-editpage-tos-summary - edit page, below summary/save button & above edittools
Initial English defaults are taken from
http://meta.wikimedia.org/wiki/Licensing_update/Implementation

Ariel's search for other copyright-related messages looks fairly clean,
so we should be good here for now as $wgRightsText/$wgRightsUrl are
already updated.

Messages are overridden using new hooks, which can also be used to switch the
messages based on title or namespace -- this may be desirable for sites such
as mediawiki.org which keep some namespaces PD etc.

Use of the overridden messages by WikimediaMessages is switched in when
$wgRightsUrl is set to 'http://creativecommons.org/licenses/by-sa/3.0/'



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@52361 dd0e9695-b195-4be7-bd10-2dea1a65a6b6