Experimentally enable HTML 5 doctype
Per wikitech-l discussion.  This also removes a <meta> tag that was
completely pointless, but theoretically required by XHTML 1, yet causes
the HTML 5 validator to complain.

Together with a few other recent commits, this should cause at least
Special:BlankPage to validate as HTML 5.  Real pages may still have some
invalid markup that should be fixed -- in a large majority of cases,
markup that was poor practice in XHTML as well.  (The Wikimedia
copyright icon needs to have border="0" removed for everything to
validate 100% on Wikipedia.)

By Brion's request, this is still behind a config option, and can be
disabled with $wgHtml5 = false;.  I expect we'll be able to remove that
option in the near future, once everyone is satisfied that changing two
lines of HTML output that all browsers have demonstrably ignored since
the Paleolithic won't cause all Wikipedia users' computers to undergo a
spontaneous uncontrolled fission reaction or something.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@53034 dd0e9695-b195-4be7-bd10-2dea1a65a6b6