Fix regression in r53665 which broke ability to act as OpenID server.
The logic was reversed when changing from a string check to array check, so we would refuse to act as an OpenID server for local users, instead of refusing to act as OpenID server for remotely-sourced users.
With this and the last fix, I'm now actually able to use a MediaWiki instance as an OpenID auth source for other sites (tested w/ StatusNet 0.9.x dev + OpenID plugin on PHP 5.2.11).


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@59041 dd0e9695-b195-4be7-bd10-2dea1a65a6b6