(bug 21387) Make $ regex work for the URLs. Patch contributed by Platonides.
Bug comment: Set PCRE_MULTILINE on spamblacklist regexes. $ on spam blacklist regex should match the end of the url (not of the text) so it can be used to match only the mainpage. Since the candidate urls are already joined with a new-line separator, it's just setting PCRE_MULTILINE on the regex.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@60869 dd0e9695-b195-4be7-bd10-2dea1a65a6b6