Delete CanonicalHref extension
This extension is broken in a few ways:

1) It outputs rel="canonical" on every single page even if it points to
a totally unrelated page.  E.g., on action=history, it outputs a
canonical link pointing to the article itself.  Canonical links are to
point to a canonical version when there are several very similar pages,
as a hint to search engines which to return in search results.  It does
*not* mean "donate all this page's PageRank to some different page".
The extension also outputs rel="canonical" pointing to the current page,
which is silly.

2) It outputs rel="canonical" even where core code also does, so there
are two identical links.  Unless for some reason they aren't always
identical.

3) If any improvements are to be made, they should be made in core.
Such a simple, useful feature should not be in an extension.

Some people are probably still using this extension in the usual
cargo-cult tradition of SEO.  Hopefully this will get some of them to
stop.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@60913 dd0e9695-b195-4be7-bd10-2dea1a65a6b6