Add disabled <datalist> support to mwsuggest
<datalist> is an HTML5 element that allows us to provide suggestions to
the browser to display with the proper format with no extra work on our
part.  I got a working implementation, so I'm committing it for the
future, but the only supporting browser (Opera) has a small but fatal
flaw, so it's disabled unconditionally.

I moved around some code to separate out the <div>-specific parts from
the general code, which unfortunately makes this somewhat messy to
review.  (I should probably have done it in two parts.)  There aren't
supposed to be any functional changes in any browser, though, so if you
find any, either fix or revert.

Before enabling this feature, we'd also want to look at UI in shipping
browsers.  WebKit looks like it will implement it more as a combobox, in
which case we wouldn't want to use it for WebKit.  But the code will
still be useful if/when this or any other search-suggest feature becomes
standard.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@61348 dd0e9695-b195-4be7-bd10-2dea1a65a6b6