Support for using Google's translation api as a source for tm suggestions and an interface element for copying tm suggestion into text area.
I suppose should use Git... Change is tested for with both the old
and the new js translation edit page. Since the js system cannot use
only one id for textarea anymore, it is now necessary to pass the
id used for textarea to translationhelpers. If javascript is not
enabled, the link will do nothing.

Google suggestions will use $wgGoogleApiKey if defined, but according
to the documentation it is not mandatory. In addition user ip is
passed to Google after salted hashing. This should be safe enough
to fullfill the use case without actually revealing any private
data. The code uses hard-coded two second timeout for now. In worst
case this leads to five second wait for the user if both services are
down. My JavaScript-Fu is not strong enough to implement async
loading and parsing on the client side.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@61365 dd0e9695-b195-4be7-bd10-2dea1a65a6b6