First part of fixing js edit which is broken with Vector skin
The core problem is that jQuery is loaded twice, once by Translate and
once by UI extension. The latter breaks ajax form plugin which is loaded
in the middle.

Changed to load core jQuery instead of custom one (which again means no
support for 1.15 and older :< ). Since the core jQuery does not include
any plugins, I added a custom set that includes the js, images and css
needed for js edit.

Changed the code to use $wgExtensionAssetsPath at one place while at it.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@61429 dd0e9695-b195-4be7-bd10-2dea1a65a6b6