Fix regression introduced in r30117: Don's sort() the author list in Special:Version
When extensions specify a list of authors they expect them to appear
in Special:Version as they're specified. Sorting lists on
Special:Version as introduced in r30117 makes sense for things like
the names of parser functions, but not authors.

Some extensions give arrays like array("foo", "bar", "others") and
expect output like "foo, bar and others". Sometimes (like in the case
of Maps.php) this would only incidentally work.

Furthermore in some cases the first author in the array indicates the
primary author, this is the case with CheckUser.php and other similar
extensions with multiple authors.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@62057 dd0e9695-b195-4be7-bd10-2dea1a65a6b6