CodeReview: Fixed error in CodeRevisionView::formatPathLine().  $safePath was being used to output the path when ViewVC is disabled, but is only being created inside the code path where it is enabled.  Have moved this outside the if statement so it is available to both code paths.
I also made the path bold so that it stands out more.  It's a lot harder to read this section if the paths are not differentiated from the normal text in some way.  Originally, I only added this for when ViewVC was disabled, as the link colouring when it is enabled is enough to make it stand out.  Then I realised that deleted paths are not made into links (for obvious reasons) and so added a bold there too, but then it looked weird to have the links in normal font and the non-links in bold, so I added it for all situations.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@62515 dd0e9695-b195-4be7-bd10-2dea1a65a6b6