Fix validity and correctness of r63342
The name attribute may only be used on <a> if an id attribute with
identical value is given on the same element.  I moved the id to the tr
so it leaves a bit of space at the top, rather than using a name
attribute at all.  Also, we need to call Sanitizer::escapeId() on the
group key, lest people use spaces or something in the group key (not
advisable, but possible).

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@63495 dd0e9695-b195-4be7-bd10-2dea1a65a6b6