(bug 23075) correct MediaTransformError default width in gallery. Contributed by Derk-Jan Hartman.
Submitter's remarks:
This patch changes the minimum size of a MediaTransformError to 120px, which
makes sure it will at least fit galleries. Any smaller and likely it would
become unreadable, so keep the ugly in those cases.

The patch also changes the message from a table to a div, because there is no
real need to use a table here. I have added "display:inline-block" to the css
to make it behave almost exactly as an img or table element would (This might
be important for [[File:image.gif|180px]], where no other framing is provided.)

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@65350 dd0e9695-b195-4be7-bd10-2dea1a65a6b6