Fixed translation completion percentage calculation bug.
This was triggered when new the "do not mark translations outdated"
switch was used. The calculation would however see that translation
was made against an older version and added 20% penalty for it.

The fix is to consider the number of changes since the version the
translation was made against only if the message is currently marked
outdated.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@68605 dd0e9695-b195-4be7-bd10-2dea1a65a6b6