new-installer: GRANT permissions to our new non-root user by sourcing users.sql
The old installer sourced maintenance/users.sql to create a new
non-root user if requested. Do that in the new installer as well.

So far this only works on MySQL, but I'm adding a generic setupUser
function on the assumption that other databases (except SQLite) want
to do some sort of user setup.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@69008 dd0e9695-b195-4be7-bd10-2dea1a65a6b6