new-installer: Delay database object construction until ->execute time
My method of putting code that alters $this->parent->installSteps in
the MySQL constructor didn't work because the installer will construct
objects for all the databases, even those it doesn't use.

This ostensibly happens because it needs to be able to provide
defaults for all of them on the DBConnect page.

But when I was going to fix that by exiting the MySQL constructior by
checking $wgDBtype I found that it didn't work, because WebInstaller
calls Installer's __construct *before* any sessions are read or set
up, so $wgDBtype will always be mysql, since that's the default.

Fix that by delaying the construction of the database objects. The
WebInstaller (or equivalent) now has to call ->setupDatabaseObjects()
in its ->execute method. This way the defaults on the DBConnect will
still be provided, but we'll have access to session data in the
database constructors.

Ughed-by: Chad <innocentkiller@gmail.com>

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@69108 dd0e9695-b195-4be7-bd10-2dea1a65a6b6