Drop OAI repository check for newSchema
IT's comparing against 1.5 alpha

And then on [[Extension:OAIRepository]]

Warning: The current version of OAI won't work with MW1.12 or lower, since the add of wfGetLB() (LBFactory abstract class) in rev:32578. To uses with 1.12, download this version of the files. Make sure you use the ExtensionDistributor by going to the "download snapshot" link in the infobox to help you get the right version.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@69995 dd0e9695-b195-4be7-bd10-2dea1a65a6b6