Fix extensions to work with category sort changes
Three different extensions were copy-pasting all of
CategoryPage::closeShowCategory() just to change which CategoryViewer
descendant was used.  I changed the base class to use a member variable
for the CategoryViewer class, so this reduces code duplication too.  I
only tested this on CategoryTree; MetavidWiki is certainly still broken,
from the looks of it (assumes $this->from is a string), but
PeopleCategories might be okay.  Hard to avoid breaking extensions with
this sort of schema/index change.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@71051 dd0e9695-b195-4be7-bd10-2dea1a65a6b6