Ensure that the first item in graphs is a whole hour, day, week or month.
Before it would just cut off right at N days before current time, go
backwards to midnight if lucky. Now it goes as many days backwards
until it finds itself at the beginning of the hour/day/week/month.
Also mark the last item with a star to remind the data for that item
is not yet complete. Any ideas how to do this better?


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@71463 dd0e9695-b195-4be7-bd10-2dea1a65a6b6