Merge hooks. JS is used to add stuff onto pages. Do restriction now in addResources
As before, only for NS_MAIN, but also, has to be in $wgArticleAssessmentPages

To add stuff... After including the extension in LocalSettings, add

$wgArticleAssessmentPages = array( 1 => 'Main page', 25 => 'Something' );

ie pageid => title

Right hand side of the => doesn't matter hugely, just neededed to force the pageid to be the key. Some maintainability i suppose also?

Remove some trailing whitespace

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@72071 dd0e9695-b195-4be7-bd10-2dea1a65a6b6