Update comment in Category.php
Per code review I gave on r72547.  I'm actually not sure how to rewrite
this so it works with all DBs -- a replace() variant that called
nextSequenceValue() if the row didn't already exist, and reused the
existing id otherwise, would perhaps be ideal.  (Speaking of which,
replace() needs to be properly factored out, with the DBMS-independent
version in DatabaseBase and the MySQL one in DatabaseMysql.)

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@72596 dd0e9695-b195-4be7-bd10-2dea1a65a6b6