Followups mostly to r75906.
* Improved logic around error messages, and understandability of those messages.

* Removed FIXME on Content-Length header. When we serve the file
this way, it seems to be necessary to explicitly set the Content-Length. I tried 
removing it and unsurprisingly it did not appear.

I am unsure what RoanKattouw is referring to when he says PHP can do this. 
How can PHP know what the Content-Length is going to be, unless we set this explicitly?




git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@76746 dd0e9695-b195-4be7-bd10-2dea1a65a6b6