Enable mwsuggest datalist implementation
Tested in Firefox 4.0b8, which is the only browser it will trigger for
due to the feature test.  Opera is explicitly blacklisted for the reason
given in the comment.  This feature was added in r61348 but remained
disabled by default because no browser actually supported it properly
(due to Opera's bug).  Of course, this is somewhat obsolescent given the
fancy new Vector search, but it's a two-line change, so why not.

I was reminded of this by Jonas Sicking, a Firefox developer, remarking
on how <datalist> should indeed be usable for this purpose in Firefox 4:
http://lists.whatwg.org/htdig.cgi/whatwg-whatwg.org/2010-December/029577.html

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@79363 dd0e9695-b195-4be7-bd10-2dea1a65a6b6