* (bug 26498) allow LinksUpdate with API
Original patch by Umherirrender. Tweaked to take account of User->pingLimiter() (ie only update links if not limited)

Possible TODO: Error if we are limited? $this->dieUsageMsg( array( 'actionthrottledtext' ) ); seems a bit harsh

Maybe just have a linkupdate="$this->parseMsg( array( 'actionthrottledtext' ) )" in the result? Or something. Definately notification of error to some respect is needed. Just dieing and stopping the whole process is OTT?


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3/includes@79638 dd0e9695-b195-4be7-bd10-2dea1a65a6b6