js email validation fixed + i18n support
The Javascript email validation in User:Preference got broken by an id change,
this patch fix the issue.

A long standing issue was a lack of proper i18n support, message helper
was hard coded in the js file.  It now use the resourceloader to get the
i18nized message.

Follow up r75670



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@79904 dd0e9695-b195-4be7-bd10-2dea1a65a6b6