Ignore code coverage for compatibility and shell functions
The compatibility functions in GlobalFunctions are just wrapper for
their equivalent in the Fallback class.  We should test the implementation
and we can safely ignore those wrappers.

Shell functions ignored make use of sleep() which is evil. They also
do some outputs to the console which is probably hard to test properly.
Given they are not critical, I just ignore their code coverage, we can
still test them though :)



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@81610 dd0e9695-b195-4be7-bd10-2dea1a65a6b6