Use the content language to set the skin direction, instead of the user language. This makes the content readable and the UI messed up, instead of vice versa.
It's not ideal, but at least it's no worse than it was in 1.16. It's preferable because the content area is more likely to have bidirectional text than the UI area. The UI is split up into short strings, generally in a consistent language, so there is less chance of the UBA moving fragments of text around. 

Ideally, the direction of each div would be set according to the language of the text it contains. 


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@81622 dd0e9695-b195-4be7-bd10-2dea1a65a6b6