* (bug 27230) Alignment fix for Vector watch tab icon spinner in IE 7
The CSS for the loading tab had:
	background-position: center 60%;

The 'center' apparently worked for other browsers, but in IE7 it was getting a bit off, and threw the image over a few pixels.
'center' keyword here is allegedly equivalent to 50% which I also tried, but with the same results.

Instead, using the same fixed position that we specify for the regular rest-state icon works:

	background-position: 5px 60%;

Resulting looks ok for me in:
* IE 6 / XP
* IE 7 / XP (fixes regression)
* IE 8 / XP
* Firefox 4b10 / Linux

Shouldn't hurt anything else.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@81679 dd0e9695-b195-4be7-bd10-2dea1a65a6b6