Rewrite the caching code in nextJobDB.php so that jobs-loop.sh has a chance of being able to exit from the "high-priority" loop.
The current syndrome is that when the cache expires, a small number of high-priority jobs are added. These jobs are quickly cleared, but the job runners do not continue on to other things, because the cache keeps returning the same small list of wikis for 5 minutes. After 5 minutes, there are a few more jobs in the high-priority list, so the cycle continues.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@83634 dd0e9695-b195-4be7-bd10-2dea1a65a6b6