setting servername with an IPv6 request must ensure we have both brackets
On lighttpd 1.4.28, the SERVER_NAME CGI variable is truncated at the first
colon. This makes it return an incorrect value for SERVER_NAME when the user
make the request to an IPv6, it outputs something like [2001.

This patch make sure we have either both opening and closing brackets or no
brackets at all (hence the 'xor' boolean check).

+ipv6



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@83847 dd0e9695-b195-4be7-bd10-2dea1a65a6b6