* updated tableName function changed addIdentifierQuotes, added isQuotedIdentifier and removeIdentifierQuotes * quoting objects in Oracle is poorly supported prior to 10g (it still has bugs in 10g) so i wish to avoid it for as long as possible * i've added /*Q*/ marker to avoid double-prefixing of table names * tableName quoted parameter is usable only in cases where you call it directly in functions where it might occur that tableName can get called twice it is unusable as the tableName gets prefixed twice
Oracle documentation is generally crappy. They claim a lot of things but some of those claims depend on a gazillion of factors and in case of one of this factors being a bit off have no general solution. Object name quoting is one of such things and is general practice amongst oracle DBAs that if you write the code directly you can use them but if you write abstracted or embedded code it's best to stay away if you can.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@86112 dd0e9695-b195-4be7-bd10-2dea1a65a6b6